Restructure fds_fximport to improve import time
Also add short-circuit frontex message detection
